// src/server/api/models/banking/models/CashTransaction.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { CashTransactionStatus } from './CashTransactionStatus';
import type { CashTransactionType } from './CashTransactionType';
import type { LocalDate } from './LocalDate';
import type { LocalDateTime } from './LocalDateTime';
import type { SubAccountType } from './SubAccountType';
import type { UUID } from './UUID';

export type CashTransaction = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The journal ID used to process the payment
     */
    journalId?: UUID | null;
    /**
     * The client ID from the crm service
     */
    clientId: UUID;
    /**
     * The client description
     */
    clientLabel: string;
    /**
     * The client code from CRM
     */
    clientCode: string;
    /**
     * The account ID
     */
    accountId: UUID;
    /**
     * The account label
     */
    accountLabel: string;
    /**
     * The account number
     */
    accountNo: string;
    /**
     * The sub account type
     */
    subAccountType: SubAccountType;
    /**
     * The teller ID
     */
    tellerId: UUID | null;
    /**
     * The ID of the branch that processed the transaction
     */
    branchId: UUID;
    /**
     * A unique code generated by the system
     */
    code: string;
    /**
     * The transaction amount
     */
    amount: number;
    /**
     * The transaction currency
     */
    currency: string;
    /**
     * The transaction description
     */
    label: string;
    /**
     * The transaction reference
     */
    reference: string;
    /**
     * The transaction notes
     */
    notes?: string | null;
    /**
     * The status of the transaction
     */
    status: CashTransactionStatus;
    /**
     * The type of the transaction
     */
    type: CashTransactionType;
    /**
     * The transaction date
     */
    transDate: LocalDate;
    /**
     * The idempotency key
     */
    idemPotencyId?: UUID | null;
};